+++ src/Sentry.AspNetCore/SentryAspNetCoreOptionsSetup.cs
@@ -3,7 +3,6 @@ using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.Options;
 using Sentry.Extensions.Logging;
 using Sentry.Internal;
-using Microsoft.Extensions.Hosting;
 #if NETSTANDARD2_0
 using IHostingEnvironment = Microsoft.AspNetCore.Hosting.IHostingEnvironment;
 #else