+++ src/Sentry.Serilog/SentrySinkExtensions.cs
@@ -4,8 +4,6 @@ using System.IO.Compression;
 using System.Linq;
 using System.Net;
 using Sentry;
-using Sentry.Internal;
-using Sentry.Protocol;
 using Sentry.Serilog;
 using Serilog.Configuration;
 using Serilog.Events;