+++ src/Sentry/Infrastructure/ConsoleDiagnosticLogger.cs
@@ -1,6 +1,5 @@
 using System;
 using Sentry.Extensibility;
-using Sentry.Protocol;
 
 namespace Sentry.Infrastructure
 {