+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## vNext
- 
+
+* test. (#645) @lucas-zimerman
 * Ref moved SentryId from namespace Sentry.Protocol to Sentry (#643) @lucas-zimerman
 * Ref renamed `CacheFlushTimeout` to `InitCacheFlushTimeout` (#638) @lucas-zimerman